refactor drivers into mail folder and add driver option to mailable
