+++ src/masonite/mail/__init__.py
@@ -1,4 +1,3 @@
 from .Mail import Mail
 from .Mailable import Mailable
-from .Welcome import Welcome
 from .MockMail import MockMail
