+++ src/masonite/mail/drivers/MailgunDriver.py
@@ -1,5 +1,5 @@
 import requests
-from .Recipient import Recipient
+from ..Recipient import Recipient
 
 
 class MailgunDriver:
