+++ src/masonite/providers/MailProvider.py
@@ -1,6 +1,6 @@
 from .Provider import Provider
 from ..mail import Mail
-from ..drivers.mail import SMTPDriver, TerminalDriver, MailgunDriver
+from ..mail.drivers import SMTPDriver, TerminalDriver, MailgunDriver
 from ..utils.structures import load
 from ..mail import MockMail
